some cleanups

+++ scripts/snet/time_check/RUN.sh
@@ -1,6 +1,10 @@
 #!/bin/sh
 
-kill -9 `cat twistd.pid`
-sleep 1
-mv twistd.log twistd.log.SAVE
+if [ -e twistd.pid ]; then
+	kill -9 `cat twistd.pid`
+	sleep 1
+fi
+if [ -e twistd.log ]; then
+	rm twistd.log
+fi
 twistd -y sample.py
